nft-test: check start of rule with sets too
We special case rules with sets ({}) so set members are sorted properly.

But we failed to check start of rule:
input:  meta mark { 1, 2 } bar
expect: meta mark { 1, 2 } bar

We made sure 'bar' is equal in both input and expected output,
but we did not check start (before {).

This causes a gazillion of warnings, will fix in followup commit.

Signed-off-by: Florian Westphal <fw@strlen.de>