src: netlink_delinarize: don't stop when encountering unsupported expression
The error message is still shown, but try to make sense of further
expressions (if any).

I tried to replace the expression by a textual representation.

Two variants I tested are:
1. append as comment:
   ip saddr 127.0.0.2 drop comment "unknown expression 'foo'"
   This allows nft -f, but it adds/alters a comment.

2. substitute in-place (i.e., add a constant expression
   with the error message instead of 'unkown' message:

   unknown expression "foo" ip saddr 127.0.0.2 drop

This won't allow 'nft -f' however.
I'm not sure this is a problem, in any case we lose information
if we can't find an expression (e.g. added by newer version).

Signed-off-by: Florian Westphal <fw@strlen.de>