tests: use 100ms for set timeout
Pablo reports set test fails with HZ=250, as it lists "324ms" instead
of "321".  This is because of rounding errors that occur when converting
from user-side millisecond scale to kernel-internal jiffies one.

use 100ms for now to avoid this error.
Alternatives would be to store use-provided value in kernel or to avoid
the conversions; this would require a change to make timeout independent from
jiffies on kernel side.

Reported-by: Pablo Neira Ayuso <pablo@netfilter.org>
Signed-off-by: Florian Westphal <fw@strlen.de>