doc: nft.txt: Wrap extra long lines to 80 chars
When I tried to send a patch that included man page update I got the
following error from git send-email:
	fatal: patch.patch:287: patch contains a line longer than 998 characters
Line 287 was a non-modified line so it was there before my patch.

Even this patch can only be sent with mutt but not with git send-email.

This patch tries to fix this issue by wrapping extra long lines to 80
characters wide.

Signed-off-by: Máté Eckl <ecklm94@gmail.com>
Signed-off-by: Florian Westphal <fw@strlen.de>