src: Make invalid chain priority error more specific
So far if invalid priority name was specified the error message referred
to the whole chain/flowtable specification:

	nft> add chain ip x h { type filter hook prerouting priority first; }
	Error: 'first' is invalid priority in this context.
	add chain ip x h { type filter hook prerouting priority first; }
	                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

With this patch this reference is made specific to the priority
specification:

	nft> add chain ip x h { type filter hook prerouting priority first; }
	Error: 'first' is invalid priority in this context.
	add chain ip x h { type filter hook prerouting priority first; }
	                                               ^^^^^^^^^^^^^^

`prio_spec` is also reused to keep naming intuitive. The parser section
formerly named `prio_spec` is renamed to `int_num` as it basically
provides the mathematical set of integer numbers.

Signed-off-by: Máté Eckl <ecklm94@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>