doc: Review man page building in Makefile.am
Previously, changes to any of the included adoc snippets in nft.txt were
not detected and hence the man page not updated (unless 'make clean' was
called). It seems like the '.txt.8' target only considers foo.txt when
trying to generate foo.8, so get rid of that and introduce a dedicated
target for nft.8.

While doing so, apply a few other minor changes:

* Although nft.8 target has to list all included adoc snippets as a
  dependency, it is sufficient to call a2x with the main one (i.e.,
  nft.txt) only.

* Keep common a2x parameters in a variable.

* Use ${A2X} everywhere and hide all calls behind ${AM_V_GEN}, not just
  the one for nft.8.

Signed-off-by: Phil Sutter <phil@nwl.cc>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>